agri-dashboard/requirements.txt
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
Pillow==10.0.0
pytest==7.4.0
python-dotenv==1.0.0
pyyaml==6.0.1
scikit-learn==1.3.0
qgis==3.28.0
opencv-python==4.8.0